--- packages/client/src/components/profileFormField/style.scss
@@ -1,53 +1,0 @@
-.form-field {
-  display: flex;
-  align-items: center;
-  gap: 20px;
-  margin: 15px 0;
-
-  .form-field__label {
-    min-width: 140px;
-    font-weight: bold;
-  }
-
-  .form-field__input {
-    flex: 1;
-    padding: 8px 12px;
-    border: 1px solid #ddd;
-    border-radius: 6px;
-
-    &:focus {
-      outline: none;
-      border-color: #007bff;
-      box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
-    }
-  }
-
-  .form-field__input-wrapper {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    flex: 1;
-  }
-
-  .form-field__error {
-    color: #dc3545;
-    font-size: 12px;
-    margin-top: 4px;
-  }
-}
-
-@media (max-width: 768px) {
-  .form-field {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 8px;
-
-    .form-field__label {
-      min-width: auto;
-    }
-
-    .form-field__input {
-      width: 100%;
-    }
-  }
-}
